feat(i18n): complete migration to new catalog-based i18n system
BREAKING CHANGE: Complete replacement of i18n system

- Replace old viper-based i18n with new catalog system
- Move catalog files from subdirectory to main i18n package
- Delete old manager, loader, and error handling code
- Implement simplified i18n.go with global message registration
- Migrate admin module to use new Message() API with embedded defaults
- Add backward compatibility for gradual module migration
- Update main.go to initialize new i18n system
- Fix all import paths and references
- Register 80+ message keys with English defaults
- Support parameterized messages with type safety
